Add files via upload
In this Project we use Data from a Survey of Data Professionals in Different Roles and Countries. The Data tells a story of what their favorite Coding language is, an average of what their salary is, how difficult it was for them to break into a Data Role, their happiness level with their current salary and lastly the most important Work/Life Balance.